Fix empty files.
