--- cypress/component/ComponentName.cy.js
@@ -1,5 +1,0 @@
-describe('ComponentName.cy.js', () => {
-  it('playground', () => {
-    // cy.mount()
-  })
-})
