--- cypress/component/Logs.cy.js
@@ -1,5 +1,0 @@
-describe('Logs.cy.js', () => {
-  it('playground', () => {
-    // cy.mount()
-  })
-})
