+++ src/Pages/Home.js
@@ -4,7 +4,7 @@ function Home() {
 	return (
 		<div className='Home'>
 			<h2>Welcome</h2>
-			<h3>To Captain's Log App! Live Long and Prosper 🖖!</h3>
+			<h3>🖖 To Captain's Log App! Live Long and Prosper 🖖!</h3>
 		</div>
 	);
 }
